Update testscript + source uninstaller

+++ uninstall.groovy
@@ -2,11 +2,11 @@
 
 import com.mongodb.*
 
-println "Groovy uninstalling >>>>>>>"
-println "uninstall_service:" + this.args[0]
-println "delete_data:" + this.args[1]
-println "delete_files:" + this.args[2]
-println "Groovy uninstalling >>>>>>>"
+println "========================="
+println "Uninstall service:" + this.args[0]
+println "Delete MongoDB data:" + this.args[1]
+println "Delete wiperdog's files:" + this.args[2]
+println "========================="
 
 def rmService = (this.args[0] == "TRUE")?true:false
 def rmMongoData = (this.args[1] == "TRUE")?true:false
@@ -15,12 +15,16 @@ def rmFiles = (this.args[2] == "TRUE")?true:false
 if(rmService){
 	uninstallService()
 }
-if(rmFiles){
-	uninstallFile()
-}
+sleep 2
 if(rmMongoData){
 	uninstallMongoData()
 }
+sleep 2
+if(rmFiles){
+	uninstallFile()
+}
+sleep 2
+println "*** WIPERDOG UNINSTALLED SUCCESSFULL !"
 def executeCommand(List<String> listCmd){
 	File workDir = new File(System.getProperty("user.dir"));
 	ProcessBuilder builder = new ProcessBuilder(listCmd);
@@ -33,7 +37,7 @@ def executeCommand(List<String> listCmd){
 	//println proc.err.text
 }
 def uninstallService(){
-	println ">>>UNINSTALL WIPERDOG SERVICE"
+	println "*** UNINSTALL WIPERDOG SERVICE..."
 	String osName = System.getProperty("os.name").toLowerCase() 
 	if(osName.indexOf("win") == -1){//-- LINUX
 		//Stop service
@@ -45,7 +49,13 @@ def uninstallService(){
 		executeCommand(listCmd)
 			
 		//Remove service
-		List<String> listCmd = new LinkedList<String>()
+		listCmd = new LinkedList<String>()
+		listCmd.add("sudo")
+		listCmd.add("rm")
+		listCmd.add("-f")
+		listCmd.add("/etc/init.d/wiperdog")			
+		executeCommand(listCmd)
+		listCmd = new LinkedList<String>()
 		listCmd.add("sudo")
 		listCmd.add("update-rc.d")
 		listCmd.add("-f")
@@ -83,14 +93,18 @@ def uninstallService(){
 		listCmd.add("delete")
 		listCmd.add("wiperdog")
 		executeCommand(listCmd)
+		println "*** WIPERDOG'S SERVICE REMOVED !"	
+
 	}
 	
 }
 
 def uninstallFile(){
-	println ">>>UNINSTALL FILES"
+	println "*** REMOVED WIPERDOG'S FILES..."
 	File WDHome = new File(System.getProperty("WIPERDOG_HOME"))
 	deleteAllFileAndFolder(WDHome)
+	println "*** WIPERDOG'S FILES REMOVED !"
+
 }
 
 def deleteAllFileAndFolder(path){
@@ -100,9 +114,7 @@ def deleteAllFileAndFolder(path){
 			if(file.isDirectory()){
 				deleteAllFileAndFolder(file)
 			}else{
-				if(!file.getName().contains("uninstall")){
-					println "Delete $file: " + file.delete()
-				}
+				file.delete()
 			}
 		}
 	}
@@ -111,7 +123,7 @@ def deleteAllFileAndFolder(path){
 }
 
 def uninstallMongoData(){
-	println ">>>UNINSTALL MONGO DATA"
+	println "*** REMOVE MONGODB DATA..."
 	// Remove data in localhost
 	Mongo mongo = new Mongo()
 	DB db = mongo.getDB('wiperdog')
@@ -128,19 +140,30 @@ def uninstallMongoData(){
 					//Check config: [mongoDB: "localhost:27017/wiperdog"] 
 					if(des.mongoDB != null){
 						try{
-							def add = des.mongoDB
+							def add 
+							def dbName 
 							if(des.mongoDB.contains("/")){
-								add = des.mongoDB.substring(0, des.mongoDB.indexOf("/"))
+								add = des.mongoDB.substring(0,des.mongoDB.indexOf("/") )
+								dbName = des.mongoDB.substring(des.mongoDB.lastIndexOf("/") + 1)
 							}
-							mongo = new Mongo(add)
-							db = mongo.getDB('wiperdog')
-							db.dropDatabase()
+
+							if(add != null && dbName != null ) {
+								mongo = new Mongo(add)
+								db = mongo.getDB(dbName)
+								db.dropDatabase()
+								println "*** WIPERDOG'S MONGODB DATA REMOVED..."
+							} else {
+								println "Failed to access to MongoDB at ${add} with database name is:  ${dbName}"
+							}
 						}catch(ex){
+							println "Failed to delete wiperdog's data from MongoDB at ${des}... Please delete manually !"
 							ex.printStackTrace()
 						}
 					}
 				}
 			}
-		}
+		}	
+	} else {
+		println " Params file not existed : ${paramFile.getAbsolutePath()}"
 	}
 }
